# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pre-market agent API for MU and SMCI monitoring
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Report {
  id: number;
  date: string;
  content: string;
  tickers: string[];
  createdAt: string;
}

export interface Observation {
  id: number;
  ticker: string;
  date: string;
  summary: string;
  sentiment: string;
  /** @nullable */
  priceAtObservation?: number | null;
  createdAt: string;
}

export interface SentimentSummary {
  ticker: string;
  bullish: number;
  bearish: number;
  neutral: number;
  lastSentiment: string;
  lastDate: string;
}

export interface AgentRunResult {
  reportId: number;
  message: string;
}

export interface AgentStatus {
  running: boolean;
  /** @nullable */
  lastRunAt: string | null;
  /** @nullable */
  currentStep?: string | null;
}

export type ListObservationsParams = {
ticker?: string;
limit?: number;
};