# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pre-market agent API for MU and SMCI monitoring
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a specific report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
  /reports/latest:
    get:
      operationId: getLatestReport
      tags: [reports]
      summary: Get today's latest report
      responses:
        "200":
          description: Latest report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: No report yet today
...
components:
  schemas:
...
    Report:
      type: object
      required: [id, date, content, tickers, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
        content:
          type: string
        tickers:
          type: array
          items:
            type: string
        createdAt:
          type: string
...