# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pre-market agent API for MU and SMCI monitoring
 * OpenAPI spec version: 0.1.0
 */

export interface Report {
  id: number;
  date: string;
  content: string;
  tickers: string[];
  createdAt: string;
}